feat: Add build step and use typescript (#2)
* Initial build setup

* Use globalIgnores to ensure not even trasversing

* Avoid ts looking into dist

* Update CI to build instead of typecheck

* Add declarations and update package export path

* Add clean and cleanbuild scripts

* Add common type definitions

* Update config to use ts inference

* Set-up linting with ts files

* Add typescript-eslint as a peer dependency

* Add prepare script

* Add postinstall script as required by yarn 2

* Fix import paths

* Copy setup from typescript-eslint/parser

* Add forgotten src

* Remove extensions from imports

* Add sourceMap

* Add yarn install as part of postinstall step

* Revert "Add yarn install as part of postinstall step"

This reverts commit 347f2f58c18182d55ceb61cb7a5e86de568c025a.

* Add remaining dist/src packages

* Revert "Add remaining dist/src packages"

This reverts commit ef7e1900e41fd1499e1bc4022c84fe2db6f01d0b.

* Libs should not use tsconfig paths option

* Remove unnecessary tsconfig options

* Set nodenext version and require file extensions

* Remove unnecessary ts-check comments

* Do not require isolatedModules since we compile with tsc

* Include tsconfig.json

* Revert "Include tsconfig.json"

This reverts commit c0a59bc679ba2d742a41fb7fb38e15713b8345a6.

* Enable again skipLibCheck to solve installation issues

* Use types directly from ESLint

* Make it prettier

* Simplify default options extraction for restrict-template-expressions

* No jumps, just copy the defaults

* Remove config and type utils

* Clearer structure

* Clarify comment
