+++ .github/workflows/ci.yml
@@ -19,4 +19,4 @@ jobs:
         cache: 'yarn'
     - run: yarn install --frozen-lockfile
     - run: yarn lint
-    - run: yarn typecheck
+    - run: yarn build
