--- index.mjs
@@ -1,6 +1,0 @@
-// @ts-check
-
-import { mindoktorRecommended } from './configs/recommended.mjs';
-
-// @ts-expect-error This is correct but conflicts with the declaration file
-export default mindoktorRecommended;
