+++ package.json
@@ -3,23 +3,36 @@
   "version": "0.0.1",
   "type": "module",
   "description": "Shared ESLint config for MinDoktor projects",
-  "main": "index.js",
   "repository": "git@github.com:mindoktor/eslint-config.git",
   "author": "Min Doktor Devs",
   "license": "MIT",
   "private": false,
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "default": "./dist/src/index.js"
+    },
+    "./package.json": "./package.json"
+  },
   "scripts": {
+    "clean": "rm -rf dist",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
-    "typecheck": "tsc --noEmit"
+    "build": "tsc",
+    "typecheck": "tsc --noEmit",
+    "cleanbuild": "yarn clean && yarn build",
+    "prepare": "yarn build",
+    "postinstall": "yarn build"
   },
   "peerDependencies": {
-    "eslint": ">=9"
+    "eslint": ">=9",
+    "typescript-eslint": ">=8"
   },
   "devDependencies": {
     "@eslint/js": "^9.32.0",
     "@types/node": "^24.2.0",
     "eslint": "^9.32.0",
+    "jiti": "^2.5.1",
     "typescript": "^5.9.2",
     "typescript-eslint": "^8.39.0"
   }
