+++ src/index.ts
@@ -1,0 +1,3 @@
+import { mindoktorRecommended } from './configs/recommended.js';
+
+export default mindoktorRecommended;
